invera_todo/requirements.txt
asgiref==3.8.0
attrs==23.2.0
backports.zoneinfo==0.2.1
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
importlib-resources==6.3.2
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pkgutil-resolve-name==1.3.10
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
typing-extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
zipp==3.18.1
